requirements.txt
google-genai==1.31.0
tomlkit==0.13.3
